Move fake token infos to TransactionRepository
