+++ data/src/test/java/io/gnosis/data/repositories/TransactionRepositoryTest.kt
@@ -2,9 +2,6 @@ package io.gnosis.data.repositories
 
 import io.gnosis.data.backend.TransactionServiceApi
 import io.gnosis.data.models.*
-import io.gnosis.data.repositories.TokenRepository.Companion.ETH_SERVICE_TOKEN_INFO
-import io.gnosis.data.repositories.TokenRepository.Companion.FAKE_ERC20_TOKEN_INFO
-import io.gnosis.data.repositories.TokenRepository.Companion.FAKE_ERC721_TOKEN_INFO
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.mockk
